replace router1 with router2
