--- packages/infra-adapters/src/Openapi/atlas-plutus.ts.bak
@@ -1,3 +1,0 @@
-export * from "./atlas-plutus/JsonSchema/index.js"
-export * from "./atlas-plutus/Schema/index.js"
-export * from "./atlas-plutus/Spec/index.js"
