--- packages/infra-adapters/src/Openapi/atlas-plutus/JsonSchema/index.ts
@@ -1,191 +1,0 @@
-/* eslint-disable @typescript-eslint/ban-types */
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
-/* eslint-disable @typescript-eslint/ban-ts-comment */
-// CREDITS:
-// https://github.com/sledorze/morphic-ts/blob/master/packages/morphic-json-schema-interpreters/src/json-schema/json-schema.ts
-
-/* eslint-disable no-prototype-builtins */
-
-export interface BaseConstructor {
-  new<T>(args: /*{} extends T ? void : */ T): T
-}
-
-// @ts-expect-error
-export const Base: BaseConstructor = class<T extends {}> {
-  constructor(args?: T) {
-    if (typeof args === "object" && args != null) {
-      const keys = Object.keys(args)
-
-      for (let i = 0; i < keys.length; i++) {
-        // @ts-expect-error
-        this[keys[i]!] = args[keys[i]!]
-      }
-    }
-  }
-}
-
-export class DescriptionSchema extends Base<{
-  description?: string
-  summary?: string
-  title?: string
-  nullable?: boolean
-}> {}
-
-export class StringSchema extends Base<
-  DescriptionSchema & {
-    minLength?: number
-    maxLength?: number
-    format?:
-      | "date-time"
-      | "date"
-      | "password"
-      | "byte"
-      | "binary"
-      | "bigint"
-      | "uuid"
-      | "email"
-    pattern?: string
-  }
-> {
-  readonly type = "string"
-}
-
-export class EnumSchema extends Base<
-  DescriptionSchema & {
-    enum: ReadonlyArray<string>
-  }
-> {
-  readonly type = "string"
-}
-
-export interface NumberEnumSchema extends DescriptionSchema {
-  type: "number"
-  enum: ReadonlyArray<number>
-}
-
-export const isEnumSchema = (x: JSONSchema): x is EnumSchema =>
-  "type" in x && x.type === "string" && Array.isArray((x as EnumSchema).enum)
-
-export class NumberSchema extends Base<
-  DescriptionSchema & {
-    multipleOf?: number
-    minimum?: number
-    exclusiveMinimum?: number
-    maximum?: number
-    exclusiveMaximum?: number
-  }
-> {
-  readonly type = "number"
-}
-
-export class BooleanSchema extends Base<DescriptionSchema> {
-  readonly type = "boolean"
-}
-
-export class ArraySchema extends Base<
-  DescriptionSchema & {
-    items: SubSchema | ReadonlyArray<SubSchema>
-    minItems?: number
-    maxItems?: number
-    description?: string
-    uniqueItems?: boolean
-  }
-> {
-  readonly type = "array"
-}
-
-export interface Ref {
-  $ref: string
-}
-
-export const Ref = ($ref: string): Ref => ({ $ref })
-
-export class ObjectSchema extends Base<
-  DescriptionSchema & {
-    required?: ReadonlyArray<string> | undefined
-    properties?: Record<string, SubSchema> | undefined
-    additionalProperties?: SubSchema | undefined
-  }
-> {
-  readonly type = "object"
-}
-
-export const isObjectSchema = (x: SubSchema): x is ObjectSchema => "type" in x && x.type === "object"
-
-export type SubSchema = JSONSchema | Ref
-
-export const SubSchema = (x: SubSchema) => x
-
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface Anything {}
-
-export const Anything: Anything = {}
-
-export class OneOfSchema extends Base<
-  DescriptionSchema & {
-    oneOf: ReadonlyArray<JSONSchema | SubSchema>
-    discriminator?: {
-      propertyName: string
-    } | undefined
-  }
-> {}
-
-export class AllOfSchema extends Base<
-  DescriptionSchema & {
-    allOf: ReadonlyArray<JSONSchema | SubSchema>
-    discriminator?: {
-      propertyName: string
-    } | undefined
-  }
-> {}
-
-export type JSONSchema =
-  | StringSchema
-  | NumberSchema
-  | BooleanSchema
-  | ArraySchema
-  | ObjectSchema
-  | OneOfSchema
-  | AllOfSchema
-  | (EnumSchema & { $schema?: string })
-  | (NumberEnumSchema & { $schema?: number })
-
-export const isTypeObject = (schema: JSONSchema | SubSchema): schema is ObjectSchema =>
-  !isTypeRef(schema)
-  && (("type" in schema && schema.type === "object")
-    || schema.hasOwnProperty("properties"))
-
-export const isTypeArray = (schema: JSONSchema | SubSchema): schema is ArraySchema =>
-  !isTypeRef(schema)
-  && "type" in schema
-  && schema.type !== undefined
-  && schema.type === "array"
-
-export const isTypeRef = (schema: JSONSchema | SubSchema): schema is Ref => schema.hasOwnProperty("$ref")
-
-export const isnotTypeRef = (schema: JSONSchema | SubSchema): schema is JSONSchema => !schema.hasOwnProperty("$ref")
-
-export const isNotPrimitive = (schema: JSONSchema | SubSchema) =>
-  isTypeObject(schema) || isTypeArray(schema) || isTypeRef(schema)
-
-export const isStringSchema = (schema: JSONSchema): schema is StringSchema =>
-  "type" in schema && schema.type === "string"
-
-export const isNumberSchema = (schema: JSONSchema): schema is NumberSchema =>
-  "type" in schema && schema.type === "number"
-
-export const isBooleanSchema = (schema: JSONSchema): schema is BooleanSchema =>
-  "type" in schema && schema.type === "boolean"
-
-export const isPrimitive = (
-  schema: JSONSchema | SubSchema
-): schema is StringSchema | NumberSchema | BooleanSchema =>
-  !isTypeRef(schema)
-  && (isStringSchema(schema) || isNumberSchema(schema) || isBooleanSchema(schema))
-
-export const isObjectOrRef = (
-  schema: JSONSchema | SubSchema
-): schema is Ref | ObjectSchema => isTypeRef(schema) || isObjectSchema(schema)
-
-export const isNamed = (schema: JSONSchema | SubSchema) =>
-  !isTypeRef(schema) && "type" in schema && schema.description !== undefined
