--- packages/infra-adapters/src/Openapi/atlas-plutus/Schema/base/index.ts
@@ -1,73 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import { Ref } from "effect-app"
-import { Context, Effect } from "effect-app"
-import type { JSONSchema, SubSchema } from "../../JsonSchema/index.js"
-import { Ref as SchemaRef } from "../../JsonSchema/index.js"
-
-export interface References {
-  ref: Ref<Map<string, SubSchema>>
-}
-
-export const References = Context.GenericTag<References>("@services/References")
-
-export class UnsupportedOperation {
-  readonly _tag = "UnsupportedOperation"
-  constructor(readonly errors: ReadonlyArray<string>) {}
-}
-
-export interface ConfigExtensionRef {
-  openapiRef?: string | undefined
-}
-
-export interface ConfigExtensionMeta {
-  openapiMeta?: any
-}
-
-export type Schema = Effect<JSONSchema | SubSchema, never, References>
-
-export const SchemaURI = "SchemaURI"
-export type SchemaURI = typeof SchemaURI
-
-export function referenced(x?: ConfigExtensionRef) {
-  return (schema: Schema): Schema => {
-    if (x && typeof x.openapiRef !== "undefined") {
-      const { openapiRef } = x
-      return Effect.gen(function*(_) {
-        const { ref } = yield* _(References)
-        const jsonSchema = yield* _(schema)
-        yield* _(Ref.update(ref, (m) => m.set(openapiRef, jsonSchema)))
-        return SchemaRef(`#/components/schemas/${openapiRef}`)
-      })
-    }
-    return schema
-  }
-}
-
-export class SchemaType<E, A> {
-  _E!: E
-  _A!: A
-  _URI!: SchemaURI
-  constructor(public Schema: Schema) {}
-}
-
-export const succeed = (_: SubSchema) => Effect.sync(() => _)
-export const dieMessage = (_: string) => Effect.die(new UnsupportedOperation([_]))
-
-export function described(description: string) {
-  return Effect.map(
-    (schema: SubSchema): SubSchema => ({
-      ...schema,
-      description
-    })
-  )
-}
-
-export function titled(title: string) {
-  return Effect.map(
-    (schema: SubSchema): SubSchema => ({
-      ...schema,
-      title
-    })
-  )
-}
