--- packages/infra-adapters/src/Openapi/atlas-plutus/Schema/index.ts
@@ -1,4 +1,0 @@
-/* eslint-disable @typescript-eslint/ban-types */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
-export * from "./base/index.js"
