--- packages/infra-adapters/src/Openapi/index.ts.bak
@@ -1,4 +1,0 @@
-// tracing: off
-
-export * from "./_openapi/index.js"
-export * from "./_schema.js"
